feat(bottle): useFetch hook
