+++ apps/bottle/src/app/create-profile/_steps/region/index.tsx
@@ -1,4 +1,6 @@
 import { Stepper } from '@/components/stepper';
+import { GET } from '@/features/server';
+import { useFetch } from '@/features/server/useFetch';
 import { Step } from '@/features/steps/StepContainer';
 import { useStep } from '@/features/steps/StepProvider';
 import { useUserAgent } from '@/features/web-view/UserAgentProvider';
@@ -11,14 +13,24 @@ import { spacingStyle } from '../MBTI/MBTIStyle.css';
 import { RegionBottomSheet } from './bottom-sheet/RegionBottomSheet';
 import { regionStyle } from './regionStyle.css';
 import { SelectInput } from './select-input/SelectInput';
-import { RegionData, useFetchRegions } from './useFetchRegion';
+
+export interface RegionData {
+  city: string;
+  state: string[];
+}
+
+export interface Regions {
+  regions: RegionData[];
+}
 
 export function Region() {
   const userAgent = useUserAgent();
   const { onNextStep } = useStep();
   const { setValue, getValue } = useCreateProfileValues();
 
-  const regionsData = useFetchRegions();
+  const regionsData = useFetch<Regions>(() =>
+    GET<Regions>(`${process.env.NEXT_PUBLIC_SERVER_BASE_URL}/api/v1/profile/choice`)
+  );
 
   const selected = getValue('region');
   const [city, setCity] = useState<string | undefined>(selected != null ? selected.city : undefined);
@@ -40,8 +52,8 @@ export function Region() {
               onClose={unmount}
               items={
                 type === 'city'
-                  ? regionsData.map(({ city }) => city)
-                  : (regionsData.find(region => region.city === city) as RegionData).state
+                  ? regionsData.regions.map(({ city }) => city)
+                  : (regionsData.regions.find(region => region.city === city) as RegionData).state
               }
             />
           )}
